requirements.txt
Flask==0.10.1
Flask-RQ==0.2
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
blinker==1.4
click==4.1
github3.py==0.9.4
gunicorn==19.3.0
itsdangerous==0.24
pbr==1.4.0
redis==2.10.3
requests==2.7.0
rollbar==0.9.12
rq==0.5.4
six==1.9.0
structlog==15.2.0
uritemplate.py==0.3.0
wsgiref==0.1.2
